Update Regression test artefacts (Marriage abroad)
This commit shows the affected/changes to regression test artefacts as a result
of the Content changes affecting all third country opposite sex  outcomes in
these CNI countries:

- albania
- algeria
- angola
- armenia
- austria
- azerbaijan
- bahrain
- bolivia
- bosnia-and-herzegovina
- bulgaria
- chile
- croatia
- cuba
- democratic-republic-of-congo
- denmark
- elsalvador
- estonia
- ethiopia
- greece
- guatemala
- honduras
- hungary
- iceland
- kosovo
- kuwait
- kyrgyzstan
- latvia
- lithuania
- luxembourg
- macedonia
- mexico
- moldova
- montenegro
- nepal
- panama
- poland
- romania
- russia
- serbia
- sudan
- sweden
- tajikistan
- tunisia
- turkmenistan
- uzbekistan
- venezuela

I have reviewed these changes and found it to be as expected.

However, the following countries were missing and hard to be added to
this branch

- algeria
- angola
- bahrain
- bolivia
- bosnia-and-herzegovina
- bulgaria
- chile
- cuba
- democratic-republic-of-congo
- el-salvador
- ethiopia
- honduras
- hungary
- moldova
- nepal
- panama
- sudan
- tajikistan
- venezuela

This helps to verify that the changes only affected the third country opposite sex  outcomes.

Already existing unit, regression and integration tests passed, giving assurance
that the aforementioned commit/changes have had no effect.