Introduce PartialTemplateRenderInterceptor class
I want to be able to annotate content rendered from partial templates. This new
class, `PartialTemplateRenderInterceptor`, provides a class method, `#[]` to
build a module which is intended to be prepended to the relevant renderer class,
`ActionView::PartialRenderer`. The `handler` passed into the method is then
invoked with the template "identifier", i.e. the absolute path to template file,
and the return value from the original `#render` method. This return value can
then be modified as required.

Unfortunately I've had to duplicate the bit of logic which derives an
`identifier` for the template from either the `@template` or `@path` instance
variables within the template. This isn't ideal, because it's not robust against
changes in the implementation of `ActionView::PartialRenderer`. However, in a
later commit I plan to introduce some kind of "integration" test which ought to
catch problems of this kind.