Allow both label and suffix_label in value questions
This fixes that `label`s and `suffix_label`s cannot exist together
within the same value question.
It is not clear if this was intentional when introduced by
5974e46048836827300850b1164f6429585443e2
but it makes sense to allow both.

This also always wraps the field in a `<label>` tag
even if there is no label.