Remove hint from postcode question
A hint would be displayed twice on a postcode question,
once within the label and once within the `_current_question` partial
where all hints are displayed.
This fixes that odd behaviour and removes it from the question.

There are currently only 3 postcode questions used
among all the Smart Answers, and none is using a hint:
* all-smart-answer-questions
* benefit-cap-calculator
* landlord-immigration-check
So, nothing visible changes for the users.

This also updates the docs about `hint`s:
* Reference to `postcode` is removed due to this commit
* Reference to `value` is removed as it is obsolete;
  there was a `hint` in the now removed `text_question`
  but never one in the `value_question`,
  see also a0235687f3d41530b2aeb99643498dd91bfa8dbd