Move title, hint and error messages into label
To be more accessible and consistent with [Elements][1],
every form element should have a label
(or legend in the case of checkboxes and radio buttons),
and hints and error messages should be part of that label.

As every 'title' is a much better candidate for labels,
this moves them from the `h2` into the label (or legend).
It also moves the hints and error messages inside
and moves the inputs outside,
while making everything look nearly like before
but also improving some spacing issues.

Negative side effects:
* This leaves `label`s and `suffix_label`s more redundant than before,
  which will be addressed at some point later.
* Every `body` is now above a label,
  which will be addressed in the next commit.

[1]: http://govuk-elements.herokuapp.com/