Add /govspeak route to fix landing page bug
The govspeak version (text) version of a page should be viewable in both the
development and integration environments by appending ".txt" to the url of the
page.

In integration, appending ".txt" to the landing throws a 404 error. This is
because the ".txt" route has not been registered in panopticon.

As a work-round I've added a new ("/govspeak") route to the smartanswers app.

The govspeak for smart-answer landing pages should be accessible by going to
smart-answer/govspeak in all environments where the govspeak is allowed to be viewed.

The "/govspeak" route has a constraint on it, so the last part of the url must
be "govspeak". This so that when questions are answered by users they will go
to the next question, and that question will be rendered as html.

If the constraint didn't exist, answering the question still redirects the user
to the next question, but that question is rendered as govspeak.

Update the instructions for viewing govspeak templates in the readme to use the
new govspeak route.