Update transit for Non-Visa-National & Oversea Territory
This commit makes logic and content changes to the transit route for Overseas
Territory and Non-Visa-National countries/passport holders with the intention
of visiting the UK on transit to a Common Travel Area (i.e CTA).

In this commit the focus is on Scenario A and C in the doc[1], which is for the
Channel Island or Isle of man and somewhere else transit option.

In some areas I didn't need to make changes for changes Non-Visa-National as it
was already as required.

I have updated the tests and made changes to the content as recommended in the
google doc [1].

[1] https://docs.google.com/document/d/12PnaBt8B797xibETodkUdNTaBTMqN5yfWbvt275XAUU/edit?ts=57bc67cd#