Update transit for Visa-National & DATV
This commit makes logic for rendering the content changes to the transit route for Visa-National & DATV
countries/passport holders with the intention of visiting the UK on transit to a Common Travel Area (i.e CTA).

In this commit the focus is on Scenario D in the doc[1], which is for the Channel Island or Isle of man
and somewhere else transit option.

I have updated the tests and made changes to the content as recommended in the
google doc [1].

[1] https://docs.google.com/document/d/12PnaBt8B797xibETodkUdNTaBTMqN5yfWbvt275XAUU/edit?ts=57bc67cd#