Create a template for marriage-abroad title in Italy
There is no single rule to infer the content title for a particular
country in marriage-abroad.

Rather than have a lot of logic in the main country template that adds
exceptions for some countries, move this logic into a specific title
partial template for a country.

While this approach will create duplicate title partials, it will make it
easier for content designers to update the title for one country without
affecting the others.

No test artefacts were affected by this refactor.