Flatten content for Sweden in marriage-abroad
Following the same approach as in Italy, this commit flattens the content
for Sweden by inlining the partial content into 18 outcomes so content
designers are able to edit them directly in Github.

Update the integration tests for Sweden to point to the generic country outcome.

Regenerate the expected results file. All of the paths for Sweden now go to the
generic country outcome.

There are a few changes in artefacts, but the reason seems to be related
to the file encoding.

The regression tests fail because an outcome node is not being exercised.
outcome_ceremonies_in_sweden_when_residing_in_sweden will be removed in a later
commit.