Get embassy address from Whitehall
There is an existing calculator that queries Whitehall for a list of embassies
and the displays the address of the embassies that performs services appropriate
to that smart answer. For marriage-abroad, only embassies that provide
`Registrations of Marriage and Civil Partnerships` services are listed.

In the outcomes for Sweden (past and present) the embassy address has always
been hard coded.

Some use the address that would be pulled from Whitehall, and others use the
main embassy address.

For the outcomes that hard code the same address as in Whitehall, it is better
to re-use the code that extracts the address. This means that should the address
change, the smart answer will always be up to date.

Affected outcomes:

* sweden/third_country/partner_british/same_sex
* sweden/third_country/partner_local/same_sex
* sweden/third_country/partner_other/same_sex
* sweden/uk/partner_british/same_sex
* sweden/uk/partner_local/same_sex
* sweden/uk/partner_other/same_sex

For the outcomes where Sweden is the ceremony country, we are waiting for
confirmation from the Department about whether to change the address. Until we
hear from them, the address will stay hard-coded.

Outcomes not being changed:

* sweden/ceremony_country/partner_british/opposite_sex
* sweden/ceremony_country/partner_british/same_sex
* sweden/ceremony_country/partner_local/opposite_sex
* sweden/ceremony_country/partner_local/same_sex
* sweden/ceremony_country/partner_other/opposite_sex
* sweden/ceremony_country/partner_other/same_sex

No outcomes have visibly changed by this refactor. I think the changes being
reported are due to encoding.