Add a rates file for pay-leave-for-parents
The values for statutory pay for parents is changing again.

The pay values affected are:
- Adoption
- Maternity
- Paternity
- Shared parental

Each year, the values for these different types of parents have been the
same.

Also, the values are currently hard-coded in 8 places across 7 partial
templates, with calculator functions to determine which financial year
we’re in, so that the right amount is displayed to the user.

Rather than add another calculator function to determine whether it’s the
2017/2018 financial year, and hard-coding another 8 values, I have decided
to add implement a `rates` file instead.

The advantage of using a rates file is:
1. The pay values only appear in one place
2. The outcome templates can be stripped of a lot of their logic
3. It might be possible to remove some of the calculator functions

Assumptions - As the rates change every year, and they are only valid
for one financial year, I will assume that the start and end dates entered
in the rates file correlate to the start and end dates for the financial
year.

The step in implementing these changes is to create the rates file and
the Data Query calculator to read it’s values.

The rates/data query will be used in the following commits.

The unit tests might be a bit basic, but as this stage they are being used
to determine that we can get the behaviour we need from the calculator.

TODO: Amend the unit tests to freeze the date.